Cleanup img texture rendering fns
