+++ src/components/HeatMap/webGLUtils.ts
@@ -1,5 +1,3 @@
-import { drawTextureFullScreen } from "./drawTextureFullScreen";
-
 // -- BUFFER SETUP --
 export const createFullScreenQuad = (
   gl: WebGLRenderingContext,
@@ -83,55 +81,4 @@ export const createWebGLContext = (
   gl.blendFunc(gl.SRC_ALPHA, gl.ONE_MINUS_SRC_ALPHA);
 
   return gl;
-};
-
-export const createTextureFromImage = (
-  gl: WebGLRenderingContext,
-  image: HTMLImageElement,
-): WebGLTexture => {
-  const texture = gl.createTexture();
-  if (!texture) throw new Error("Failed to create texture");
-
-  gl.bindTexture(gl.TEXTURE_2D, texture);
-  gl.pixelStorei(gl.UNPACK_FLIP_Y_WEBGL, true); // Optional: flip Y if needed
-  gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, image);
-  gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR);
-  gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
-  gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
-
-  return texture;
-};
-
-export const createTextureFromImageSrc = async (
-  gl: WebGLRenderingContext,
-  src: string,
-): Promise<WebGLTexture> => {
-  const image = new Image();
-  image.crossOrigin = "anonymous"; // For loading cross-origin images
-  image.src = src;
-
-  await new Promise<void>((resolve, reject) => {
-    image.onload = () => resolve();
-    image.onerror = () => reject(new Error(`Failed to load image: ${src}`));
-  });
-
-  return createTextureFromImage(gl, image);
-};
-
-export const createBackgroundRenderer = (gl: WebGLRenderingContext) => {
-  let cachedTexture: WebGLTexture | null = null;
-  let cachedSrc: string | null = null;
-
-  const draw = async (src: string) => {
-    if (cachedSrc !== src) {
-      cachedTexture = await createTextureFromImageSrc(gl, src);
-      cachedSrc = src;
-    }
-
-    if (cachedTexture) {
-      drawTextureFullScreen(gl, cachedTexture);
-    }
-  };
-
-  return { draw };
 };
